# backend/tests/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Strom API",
    "description": "REST and WebSocket API for the Strom pipeline engine",
    "contact": {
      "name": "Per Enstedt",
      "email": "per.enstedt@eyevinn.se"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.3.26"
  },
  "paths": {
...
    "/api/discovery/devices": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "List all discovered devices.",
        "operationId": "list_devices",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of discovered devices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/discovery/devices/{id}": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Get a specific device by ID.",
        "operationId": "get_device",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Device ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Device not found"
          }
        }
      }
    },
...
    "/api/discovery/streams/{id}": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Get a specific discovered stream by ID.",
        "operationId": "get_stream",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Stream ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stream details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveredStreamResponse"
                }
              }
            }
          },
          "404": {
            "description": "Stream not found"
          }
        }
      }
    },
    "/api/discovery/streams/{id}/sdp": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Get the raw SDP for a discovered stream.",
        "operationId": "get_stream_sdp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Stream ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SDP content",
            "content": {
              "application/sdp": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Stream not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceCategory": {
        "type": "string",
        "description": "Device category for filtering.",
        "enum": [
          "audiosource",
          "audiosink",
          "videosource",
          "networksource",
          "other"
        ]
      },
...
      "DeviceResponse": {
        "type": "object",
        "description": "API response for a discovered device.",
        "required": [
          "id",
          "name",
          "device_class",
          "category",
          "provider",
          "properties",
          "first_seen_secs_ago",
          "last_seen_secs_ago"
        ],
        "properties": {
          "category": {
            "$ref": "#/components/schemas/DeviceCategory",
            "description": "Device category."
          },
          "device_class": {
            "type": "string",
            "description": "Device class (e.g., \"Audio/Source\", \"Video/Source\", \"Source/Network\")."
          },
          "first_seen_secs_ago": {
            "type": "integer",
            "format": "int64",
            "description": "Seconds since first discovery.",
            "minimum": 0
          },
          "id": {
            "type": "string",
            "description": "Unique ID for this device."
          },
          "last_seen_secs_ago": {
            "type": "integer",
            "format": "int64",
            "description": "Seconds since last seen.",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "description": "Display name of the device."
          },
          "properties": {
            "type": "object",
            "description": "Additional properties from the device.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "provider": {
            "type": "string",
            "description": "Provider that discovered this device."
          }
        }
      },
      "DiscoveredStreamResponse": {
        "type": "object",
        "description": "API response for a discovered AES67 stream.",
        "required": [
          "id",
          "name",
          "source",
          "multicast_address",
          "port",
          "channels",
          "sample_rate",
          "encoding",
          "origin_host",
          "first_seen_secs_ago",
          "last_seen_secs_ago",
          "ttl_secs"
        ],
        "properties": {
          "channels": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "encoding": {
            "type": "string"
          },
          "first_seen_secs_ago": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "id": {
            "type": "string"
          },
          "last_seen_secs_ago": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "multicast_address": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "origin_host": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "received_on_interface": {
            "type": [
              "string",
              "null"
            ],
            "description": "Network interface the stream was discovered on (for SAP)."
          },
          "sample_rate": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "source": {
            "type": "string"
          },
          "ttl_secs": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}